chore: update base package to dev.araopj and move salary entities to model subpackage
