+++ src/main/java/dev/araopj/hrplatformapi/salary/model/SalaryGrade.java
@@ -1,12 +1,6 @@
-package com.araopj.hrplatformapi.salary;
+package dev.araopj.hrplatformapi.salary.model;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -14,6 +8,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 @Entity(name = "salary_grade")
 @NoArgsConstructor
